+++ src/dftbp/dftbplus/initprogram.F90
@@ -1272,7 +1272,7 @@ contains
     type(TCoulombInput), allocatable :: coulombInput
     type(TPoissonInput), allocatable :: poissonInput
 
-    logical :: tInitialized, tGeoOptRequiresEgy, isOnsiteCorrected
+    logical :: tGeoOptRequiresEgy, isOnsiteCorrected
     type(TStatus) :: errStatus
 
     @:ASSERT(input%tInitialized)
@@ -1556,8 +1556,8 @@ contains
     call initElectronFilling_(input, this%nSpin, this%Ef, this%iDistribFn, this%tempElec,&
         & this%tFixEf, this%tSetFillingTemp, this%tFillKSep)
 
-    call ensureSolverCompatibility(input%ctrl%solver%iSolver, this%tSpin, this%kPoint,&
-        & input%ctrl%parallelOpts, nIndepHam, this%tempElec)
+    call ensureSolverCompatibility(input%ctrl%solver%iSolver, this%kPoint, input%ctrl%parallelOpts,&
+        & nIndepHam, this%tempElec)
     call getBufferedCholesky_(this%tRealHS, this%parallelKS%nLocalKS, nBufferedCholesky)
     call TElectronicSolver_init(this%electronicSolver, input%ctrl%solver%iSolver, nBufferedCholesky)
 
@@ -1573,9 +1573,9 @@ contains
       call initUploadArrays_(input%transpar, this%orb, this%nSpin, this%tMixBlockCharges,&
           & this%shiftPerLUp, this%chargeUp, this%blockUp)
     end if
-    call initTransport_(env, input, this%electronicSolver, this%nSpin, this%tempElec, this%tNegf,&
-        & this%isAContactCalc, this%mu, this%negfInt, this%ginfo, this%transpar, this%writeTunn,&
-        & this%tWriteLDOS, this%regionLabelLDOS)
+    call initTransport_(this, env, input, this%electronicSolver, this%nSpin, this%tempElec,&
+        & this%tNegf, this%isAContactCalc, this%mu, this%negfInt, this%ginfo, this%transpar,&
+        & this%writeTunn, this%tWriteLDOS, this%regionLabelLDOS)
   #:else
     this%tTunn = .false.
     this%tLocalCurrents = .false.
@@ -1587,6 +1587,7 @@ contains
     this%tPoisson = input%ctrl%tPoisson .and. this%tSccCalc
     this%updateSccAfterDiag = input%ctrl%updateSccAfterDiag
 
+    this%tExtChrg = .false.
     if (this%tSccCalc .and. .not.allocated(this%tblite)) then
       call initShortGammaDamping_(input%ctrl, this%speciesMass, shortGammaDamp)
       if (this%tPoisson) then
@@ -1595,17 +1596,17 @@ contains
               & this%tPeriodic, this%latVec, this%orb, hubbU, poissonInput, this%shiftPerLUp)
         #:endblock
       else
-        call initShortGammaInput_(this%orb, input%ctrl, this%speciesName, this%speciesMass,&
-            & this%uniqHubbU, shortGammaDamp, shortGammaInput)
+        call initShortGammaInput_(input%ctrl, this%speciesMass, this%uniqHubbU, shortGammaDamp,&
+            & shortGammaInput)
         call initCoulombInput_(env, input%ctrl%ewaldAlpha, input%ctrl%tolEwald,&
-            & this%boundaryCond%iBoundaryCondition, this%nAtom, coulombInput)
+            & this%boundaryCond%iBoundaryCondition, coulombInput)
       end if
       call initSccCalculator_(env, this%orb, input%ctrl, this%boundaryCond%iBoundaryCondition,&
           & coulombInput, shortGammaInput, poissonInput, this%scc)
 
       ! Stress calculation does not work if external charges are involved
       this%nExtChrg = input%ctrl%nExtChrg
-      this%tExtChrg =  this%nExtChrg > 0
+      this%tExtChrg = this%nExtChrg > 0
       this%tStress = this%tStress .and. .not. this%tExtChrg
 
       ! Longest cut-off including the softening part of gamma
@@ -1851,7 +1852,7 @@ contains
      end if
      if (tRequireDerivator) then
       select case(input%ctrl%iDerivMethod)
-      case (1)
+      case (diffTypes%finiteDiff)
         ! set step size from input
         if (input%ctrl%deriv1stDelta < epsilon(1.0_dp)) then
           write(tmpStr, "(A,E12.4)") 'Too small value for finite difference step :',&
@@ -1859,7 +1860,7 @@ contains
           call error(tmpStr)
         end if
         call NonSccDiff_init(this%nonSccDeriv, diffTypes%finiteDiff, input%ctrl%deriv1stDelta)
-      case (2)
+      case (diffTypes%richardson)
         call NonSccDiff_init(this%nonSccDeriv, diffTypes%richardson)
       end select
     end if
@@ -2363,7 +2364,7 @@ contains
       isOnsiteCorrected = allocated(this%onSiteElements)
       call ensureLinRespConditions(this%tSccCalc, this%t3rd .or. this%t3rdFull, this%tRealHS,&
           & this%tPeriodic, this%tCasidaForces, this%solvation, this%isRS_LinResp, this%nSpin,&
-          & this%tSpin, this%tHelical, this%tSpinOrbit, allocated(this%dftbU), this%tempElec,&
+          & this%tHelical, this%tSpinOrbit, allocated(this%dftbU), this%tempElec,&
           & isOnsiteCorrected, input)
 
       ! Hubbard U and spin constants for excitations (W only needed for triplet/spin polarised)
@@ -3945,19 +3946,18 @@ contains
       if (this%tFixEf .or. this%tSkipChrgChecksum) then
         ! do not check charge or magnetisation from file
         call initQFromFile(this%qInput, fCharges, this%tReadChrgAscii, this%orb, this%qBlockIn,&
-            & this%qiBlockIn, this%deltaRhoIn, this%nAtom, multipoles=this%multipoleInp)
+            & this%qiBlockIn, this%deltaRhoIn, multipoles=this%multipoleInp)
       else
         ! check number of electrons in file
         if (this%nSpin /= 2) then
           call initQFromFile(this%qInput, fCharges, this%tReadChrgAscii, this%orb, this%qBlockIn,&
-              & this%qiBlockIn, this%deltaRhoIn, this%nAtom, nEl = sum(this%nEl),&
+              & this%qiBlockIn, this%deltaRhoIn, nEl = sum(this%nEl),&
               & multipoles=this%multipoleInp)
         else
           ! check magnetisation in addition
           call initQFromFile(this%qInput, fCharges, this%tReadChrgAscii, this%orb, this%qBlockIn,&
-              & this%qiBlockIn, this%deltaRhoIn, this%nAtom,&
-              & nEl = sum(this%nEl), magnetisation=this%nEl(1)-this%nEl(2),&
-              & multipoles=this%multipoleInp)
+              & this%qiBlockIn, this%deltaRhoIn, nEl = sum(this%nEl),&
+              & magnetisation=this%nEl(1)-this%nEl(2), multipoles=this%multipoleInp)
         end if
       end if
 
@@ -4429,20 +4429,53 @@ contains
 #:if WITH_TRANSPORT
 
   !> Initialise a transport calculation
-  subroutine initTransport_(env, input, electronicSolver, nSpin, tempElec, tNegf, isAContactCalc,&
-      & mu, negfInt, ginfo, transpar, writeTunn, tWriteLDOS, regionLabelLDOS)
+  subroutine initTransport_(this, env, input, electronicSolver, nSpin, tempElec, tNegf,&
+      & isAContactCalc, mu, negfInt, ginfo, transpar, writeTunn, tWriteLDOS, regionLabelLDOS)
+
+    !> Instance
+    class(TDftbPlusMain), intent(in) :: this
+
+    !> Environment settings
     type(TEnvironment), intent(inout) :: env
+
+    !> Holds the parsed input data.
     type(TInputData), intent(in) :: input
+
+    !> electronic solver for the system
     type(TElectronicSolver), intent(inout) :: electronicSolver
+
+    !> Number of spin components, 1 is unpolarised, 2 is polarised, 4 is noncolinear / spin-orbit
     integer, intent(in) :: nSpin
+
+    !> electron temperature
     real(dp), intent(in) :: tempElec
+
+    !> Transport interface
     logical, intent(in) :: tNegf
+
+    !> Whether a contact Hamiltonian is being computed and stored
     logical, intent(out) :: isAContactCalc
+
+    !> Holds spin-dependent electrochemical potentials of contacts
+    !> This is because libNEGF is not spin-aware
     real(dp), allocatable, intent(out) :: mu(:,:)
+
+    !> Transport interface
     type(TNegfInt), intent(out) :: negfInt
+
+    !> container for data needed by libNEGF
     type(TNegfInfo), intent(out) :: ginfo
+
+    !> Transport calculation parameters
     type(TTransPar), intent(out) :: transpar
-    logical, intent(out) :: writeTunn, tWriteLDOS
+
+    !> Should tunnelling be written out
+    logical, intent(out) :: writeTunn
+
+    !> Should local density of states be written out
+    logical, intent(out) :: tWriteLDOS
+
+    !> Labels for different regions for DOS output
     character(lc), allocatable, intent(out) :: regionLabelLDOS(:)
 
     logical :: tAtomsOutside
@@ -4450,7 +4483,7 @@ contains
     integer :: nSpinChannels, iCont, jCont
     real(dp) :: mu1, mu2
 
-    ! Its a contact calculation in the case that some contact is computed
+    ! It is a contact calculation in the case that some contact is computed
     isAContactCalc = (input%transpar%taskContInd /= 0)
 
     if (nSpin <= 2) then
@@ -4501,7 +4534,7 @@ contains
 
       ! Some checks and initialization of GDFTB/NEGF
       call TNegfInt_init(negfInt, input%transpar, env, input%ginfo%greendens,&
-          & input%ginfo%tundos, tempElec)
+          & input%ginfo%tundos, tempElec, this%coord0, this%cutOff%skCutoff, this%tPeriodic)
 
       ginfo = input%ginfo
 
@@ -4897,10 +4930,10 @@ contains
     !> parallel environment
     type(TEnvironment), intent(in) :: env
 
-    !> Number of matrix rows
+    !> Size of matrix row blocks
     integer, intent(in) :: rowBlock
 
-    !> Number of matrix columns
+    !> Size of matrix column blocks
     integer, intent(in) :: colBlock
 
     !> Descriptor of the dense matrix
@@ -4994,13 +5027,10 @@ contains
 
 
   !> Check for compatibility between requested electronic solver and features of the calculation
-  subroutine ensureSolverCompatibility(iSolver, tSpin, kPoints, parallelOpts, nIndepHam, tempElec)
+  subroutine ensureSolverCompatibility(iSolver, kPoints, parallelOpts, nIndepHam, tempElec)
 
     !> Solver number (see dftbp_elecsolvers_elecsolvertypes)
     integer, intent(in) :: iSolver
-
-    !> Is this a spin polarised calculation
-    logical, intent(in) :: tSpin
 
     !> Set of k-points used in calculation (or [0,0,0] if molecular)
     real(dp), intent(in) :: kPoints(:,:)
@@ -5236,8 +5266,7 @@ contains
   !> Stop if linear response module can not be invoked due to unimplemented combinations of
   !> features.
   subroutine ensureLinRespConditions(tSccCalc, t3rd, tRealHS, tPeriodic, tCasidaForces, solvation,&
-      & isRS_LinResp, nSpin, tSpin, tHelical, tSpinOrbit, isDftbU, tempElec, isOnsiteCorrected,&
-      & input)
+      & isRS_LinResp, nSpin, tHelical, tSpinOrbit, isDftbU, tempElec, isOnsiteCorrected, input)
 
     !> Is the calculation SCC?
     logical, intent(in) :: tSccCalc
@@ -5262,9 +5291,6 @@ contains
 
     !> Number of spin components, 1 is unpolarised, 2 is polarised, 4 is noncolinear / spin-orbit
     integer, intent(in) :: nSpin
-
-    !> Is this a spin polarised calculation
-    logical, intent(in) :: tSpin
 
     !> If the calculation is helical geometry
     logical :: tHelical
@@ -5935,11 +5961,9 @@ contains
 
 
   ! Initializes short gamma calculator
-  subroutine initShortGammaInput_(orb, ctrl, speciesNames, speciesMass, uniqHubbU, shortGammaDamp,&
+  subroutine initShortGammaInput_(ctrl, speciesMass, uniqHubbU, shortGammaDamp,&
       & shortGammaInp)
-    type(TOrbitals), intent(in) :: orb
     type(TControl), intent(in) :: ctrl
-    character(*), intent(in) :: speciesNames(:)
     real(dp), intent(in) :: speciesMass(:)
     type(TUniqueHubbard), intent(in) :: uniqHubbU
     type(TShortGammaDamp), intent(in) :: shortGammaDamp
@@ -5962,12 +5986,11 @@ contains
 
 
   ! Initializes a Coulomb-calculator
-  subroutine initCoulombInput_(env, ewaldAlpha, tolEwald, boundaryCond, nAtom, coulombInput)
+  subroutine initCoulombInput_(env, ewaldAlpha, tolEwald, boundaryCond, coulombInput)
     type(TEnvironment), intent(in) :: env
     real(dp), intent(in) :: ewaldAlpha
     real(dp), intent(in) :: tolEwald
     integer, intent(in) :: boundaryCond
-    integer, intent(in) :: nAtom
     type(TCoulombInput), allocatable, intent(out) :: coulombInput
 
     allocate(coulombInput)